+++ HttpShare/Constants.cs
@@ -1,0 +1,6 @@
+namespace HttpShare;
+
+public static class Constants
+{
+	public const string LoggedInUsersOnlyPolicy = "LoggedInUsersOnly";
+}