+++ HttpShare/Controllers/UserController.cs
@@ -1,4 +1,3 @@
-using HttpShare.Models;
 using HttpShare.Sessions;
 
 using Microsoft.AspNetCore.Mvc;
@@ -10,21 +9,36 @@ public sealed class UserController(ServerSession serverSession) : CustomControll
 {
 	public const string ErrorsKey = "Errors";
 
+	[HttpGet]
+	[Route("/LogIn/")]
+	public IActionResult LogIn()
+	{
+		if (!string.IsNullOrEmpty(ServerSession.Password) ||
+			!string.IsNullOrWhiteSpace(ServerSession.Password)) return View("../User/LogInPassword");
+
+		else return View("../User/LogIn");
+	}
+
 	[HttpPost]
 	[Route("/LogIn/")]
-	public async Task<IActionResult> UserLogIn([FromForm] LogInModel logInModel)
+	public async Task<IActionResult> LogIn([FromForm] string displayName, [FromForm] string password)
 	{
-		if (ServerSession.Password == null) return RedirectPermanent("/");
-
-		if (logInModel.Password != ServerSession.Password)
+		if (!string.IsNullOrEmpty(ServerSession.Password)
+			|| !string.IsNullOrWhiteSpace(ServerSession.Password))
 		{
-			string[] errorMessages = ["Invalid password."];
+			if (!password.Equals(ServerSession.Password))
+			{
+				string[] errorMessages = ["Invalid password."];
 
-			ViewData[ErrorsKey] = errorMessages;
-			return View("Password");
+				ViewData[ErrorsKey] = errorMessages;
+				return View("../User/LogInPassword");
+			}
 		}
 
-		await UserLogIn(logInModel.DisplayName);
-		return RedirectPermanent("/");
+		string message = $"Client log in: {displayName}";
+		ServerSession.InvokeServerEvent(new ServerSessionEvent(ServerEventType.Information, message));
+
+		await UserLogIn(displayName);
+		return Redirect("/");
 	}
 }