+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## Protoype 3
+## Prototype 4
+
+- Added the "Log" tab in the advanced window
+- Added error window
+- Added mandatory display name input
+- Bug fixes
+- Code cleanup
+
+## Prototype 3
 
 - Added "Server Options" control
 - Added server port input feature